DRG: port debuffs to typescript
